Remove SB_PHP environment variable
The idea of this was to establish a convention for remote execution of
PHP scripts. But we don't really need this convention right now, it's
easier to let Score do its own thing. MediaWiki has $wgPhpCli which we
can pass through. PHP_BINDIR does not correctly return the location of
the PHP CLI binary when the binary has been relocated after build, for
example in XAMPP. So Shellbox cannot reliably detect the location of the
PHP binary, configuration is required.

Change-Id: I3cf2d53b18b9d2d2f8a137f9a432b25c989c3221