Ensure test server runs with xdebug enabled
For performance reasons, the CI containers do not have xdebug enabled by
default, it needs to be explicitly invoked with `-d zend_extension.so`.
This was not being passed along to the Shellbox server, causing
exceptions when invoking the PHPUnit coverage code.

If the main test process has xdebug loaded, pass the right parameter
along to the Shellbox server. If xdebug is already enabled by default
then it won't have any adverse effects.

Bug: T270656
Change-Id: Icde5105a1674943c763d476283a99e4749682fa4