Don't pass through the fake CGI environment to subprocesses
I was trying to be backwards compatible by passing through getenv(), but
the old code effectively only did that in CLI mode. In FastCGI mode
passthru('env') will show you a near-empty environment whereas getenv()
contains a lot of junk, even an array causing an array-to-string
conversion prior to this fix.

So, only pass through getenv() in CLI mode.

In the server modes, pass through PATH from the web server. This is a
change compared to the old MW code, but I think it's convenient.

Change-Id: Ifbeea79657c6a19644c7c038e812243db54f7810