Fix OOM in MultipartAction when handling a large request body
Guzzle's Utils::streamFor() copies php://input to a temporary file, to
fix a problem with using such a resource as the body of a subsequent
non-curl client request. https://github.com/guzzle/guzzle/issues/2276

We don't want or need that workaround since we're consuming the stream
ourselves, not passing it back to Guzzle. Upstream have noticed the
potential for an OOM here and have merged PR #463 to avoid it, but
their solution still requires unnecessary copying of the stream.

So, just use "new Stream()".

Bug: T292322
Change-Id: I10e5060daa6ea43dee430e8ef9ccdff6ceb7fd96