syntaxhighlight: add /srv/app/pygmentize script
Add a `/srv/app/pygmentize` which in turn execs the
`/usr/bin/pygmentize` script that is installed via apt. This new script
is meant to be used as an intermediate step towards a future container
where we use pip to install Pygments and are able to expose that runtime
as `/srv/app/pygmentize` but not as `/usr/bin/pygmentize` due to tooling
constraints.

This intermediate state can be used to publish a container image which
will respond identically to Shellbox calls to both the current standard
of `/usr/bin/pygmentize` and the future standard of
`/srv/app/pygmentize`. This dual path support will allow adjusting the
Wikimedia production cluster's settings to use `$wgPygmentizePath
= '/srv/app/pygmentize';` at any point after the container is deployed.

Bug: T320848
Change-Id: I7556103aebc65c376b804c2571dcb9010d42034d