Use an explicit shell inside firejail
Firejail 0.9.72 switched from using bash as the default shell to using
the user's login shell, which for a system account is typically
non-functional.

Simply prefixing the unquoted command with firejail risks a sandbox
escape because there will be a shell outside firejail in which pipes and
expansions are done.

So, run /bin/sh explicitly inside firejail, passing the complete command
as a string to this inner shell.

In the case of --seccomp=execve, it's necessary to split the command
string into literal arguments and to requote them for the outer shell.
The arguments will then be passed down to firejail and then directly to
the command. Move the SyntaxInfo memoization from Validator to Command
to support this.

Make BashWrapper be always priority 60, outside firejail, so that it will work
with --seccomp=execve. Previously --seccomp=execve was completely broken
in MW due to this issue. Nothing uses it in codesearch.

Remove --shell=none, because it causes a fatal error in Firejail
0.9.72+. It's not practical to do a version check, and the
--seccomp=execve mode was broken anyway, so this probably won't break
any real-world deployments.

Bug: T353194
Change-Id: I6f60db4db77ea2f81913db95753e3d84de97a793