dev: Allow overriding the local httpd server port
Allow local changes to the port that the Shellbox httpd is bound to via
a new SHELLBOX_HTTP_PORT environment variable. The variable defaults to
the prior 8080 port, but can now be easily overridden as needed via
a .env file.

.gitignore is also updated to ignore any local
docker-compose.override.yml configuration file. This is another means of
customizing the default docker-compose environment. See
<https://docs.docker.com/compose/multiple-compose-files/merge/> for more
details on using a compose.override.yml file.

Change-Id: I12740c3567c0c59b2e2c3152e90190ae4f47475e