Add remote download/upload support for large file performance
Allow Shellbox to download the input files for a command from a URL, and
upload the output files by a PUT request to a URL.

The idea is to use this in combination with Swift TempURL to authorize
Shellbox to write to a specific output file without giving it any stored
secrets.

I added the boolean option allowUrlFiles to the client and server. If
it is false, "URL files" (uploads and downloads) are rejected. I
considered a client-side fallback, but left it out due to YAGNI, since
MediaWiki's Swift client is a better place to do uploads in that case.

The local mode is able to upload and download if an HTTP client is
injected. For MediaWiki we probably won't inject one.

The new BoxedCommand methods inputFileFromUrl(), outputFileToUrl() and
outputGlobToUrl() are added by analogy with existing code.

To allow some abstraction of the upload method in MW's FileBackend, we
now have BoxedCommand::inputFile(), outputFile() and outputGlob(), with
factory methods for the InputFile, OutputFile and OutputGlob class
hierarchies. Configure options on files using chainable mutators on
these newly-exposed classes.

Also:
* Delete OutputEntity. On review, I found it to be poorly justified.
* Clean up LSP issues by adding OutputFileWithContents and
  InputFileWithContents. Previously InputFilePlaceholder and
  OutputFilePlaceholder required "unimplemented method" exceptions, but
  this is now fixed.
* Add userData() methods on files, to allow Shellbox to keep FileBackend
  temporary files alive by retaining references to them.
* Add onReceived() to OutputFile, allowing the user to specify a
  callback to be called when a file is received. This allows FileBackend
  to upload the file to an arbitrary store, hiding the implementation of
  the upload from the caller.

Bug: T292322
Change-Id: Ic699564572d0e83a7fd5c982e5b0e4b11a273403