Pass pcov options to child process
The test server is spawned using `php -S` and assumes a code coverage
driver to be loaded which is not always the case. On our CI, those
extensions are not loaded and are enabled solely for the coverage jobs
using command line arguments:

  php -d extension=pcov.so -d pcov.enable=1 \
    ./vendor/bin/phpunit --coverage-html /log/coverage

Those are not passed to the built-in PHP server when it is started and
it thus fails with:

  PHP Fatal error:  Uncaught RuntimeException: No code coverage driver
                    available in /src/tests/TestServer.php:74
  Stack trace:
  #0 /src/tests/TestServer.php(37): Shellbox\Tests\TestServer::selectDriver()
  #1 /tmp/sb-test-server-89ce3392e10a8780/sbtest.php(3):
       Shellbox\Tests\TestServer::main()
  #2 {main}
    thrown in /src/tests/TestServer.php on line 74

We had a similar issue with XDebug (81cf75a9f2d57). This copies the
logic for PCOV.

I have confirmed it fixes the coverage when using the CI image
docker-registry.wikimedia.org/releng/composer-php81.

Bug: T378423
Change-Id: Ic798f7662f7f2c29dda0e3e2dd25c06758300c29