Check that error level should be handled before throwing
Changes `handleError` to ignore (and not throw a ShellboxError) when
the error's level is not included in `error_reporting`. This allows
errors to be suppressed with an error control operator as documented
here:
https://www.php.net/manual/en/language.operators.errorcontrol.php

This is part of a fix for an issue where `preg_match` results in
logspam when called with an invalid regex. Since it's checking
user-created regexes, it is expected that some of these will be
invalid. `preg_match` returns `false` in these cases, but unfortunately
also emits an `E_WARNING`, which gets picked up by `handleError`.

Bug: T380751
Change-Id: I7ef612b94aeb1dd33db0e281be5e66f43f409560