Revert "Catch HTTP client exceptions and rethrow as ShellboxError"
Client objects are supplied by the caller. It is the caller's
responsibility to catch any exceptions that they throw.

The exception requirements for client objects and the fact that
exceptions are propagated back to the caller is documented on line 37 of
Client.php.

In MediaWiki in general, wrapping of exceptions is discouraged. It is
easier for admins to analyse the logs if exceptions of a certain type
have a consistent log message and backtrace. From the system
administration point of view, there's no difference between, say, a DNS
name resolution error in Shellbox and in any other layer.

You can see this in the referenced bug. A server is broken, returning an
empty response, and SyntaxHighlight would like to move that error from
the exception channel to wfWarn() which if I understand correctly is
thrown away in production.

If SyntaxHighlight really needs to do that, it can catch
ClientExceptionInterface exceptions itself, but then it needs to take on
the responsibility of proper logging.

This reverts commit e7abc6934133305f03c898b62ad86db5b699aaf1.

Bug: T374117
Change-Id: I722ef4d6c33bfa25c025dbc30bacf8402a4dd200