Add PHP 8.3 image variants
This is a straightforward copy of the existing 8.1 variants, with only
the base image overridden. No additional changes are required, largely
due to the fact that both base images are Debian bullseye.

Alas, blubber does not provide a way to parameterize variants over base
images, so this is fairly verbose. To mitigate that, I've split variants
up into sections by PHP version dependency.

Bug: T403283
Change-Id: Ie60adda269556194e481c7b983a0fdd8d0347877